fix(user-post):handle displaying post by status with user

+++ src/pages/game-details/game-posts/post-card.tsx
@@ -27,7 +27,7 @@ import { Link } from "react-router-dom";
 interface PostCardProps {
   post: GamePost;
   onViewPostDetail?: (postId: string) => void;
-  onDelete: (postId: string) => void;
+  onDelete?: (postId: string) => void;
 }
 
 const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
@@ -44,10 +44,7 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
     return post.postImages.map((image) => image.image);
   }, [post]);
 
-  const slides = useMemo(
-    () => images.map((image) => ({ src: image })),
-    [images]
-  );
+  const slides = useMemo(() => images.map((image) => ({ src: image })), [images]);
 
   const moreOptionItems: MenuProps["items"] = useMemo(() => {
     const items: MenuProps["items"] = [
@@ -68,7 +65,9 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
         key: "delete",
         icon: <FaTrash />,
         onClick: () => {
-          onDelete(post.id);
+          if (onDelete) {
+            onDelete(post.id);
+          }
         },
         danger: true,
       });
@@ -127,12 +126,7 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
 
   return (
     <div>
-      <Lightbox
-        index={currentImage}
-        slides={slides}
-        open={lightboxIndex >= 0}
-        close={() => setLightboxIndex(-1)}
-      />
+      <Lightbox index={currentImage} slides={slides} open={lightboxIndex >= 0} close={() => setLightboxIndex(-1)} />
       <div className="bg-zinc-800 w-full p-3 rounded border border-zinc-700 hover:border-orange-500 duration-300">
         <div className="flex justify-between items-center gap-3">
           <div className="flex items-center gap-3">
@@ -144,9 +138,7 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
                 <div className="font-semibold">{post.user.userName}</div>
               </Link>
 
-              <div className="text-xs text-gray-400">
-                {timeAgo(post.createdAt)}
-              </div>
+              <div className="text-xs text-gray-400">{timeAgo(post.createdAt)}</div>
             </div>
           </div>
           <Dropdown menu={{ items: moreOptionItems }} trigger={["click"]}>
@@ -155,18 +147,12 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
         </div>
 
         <div className="mt-2">
-          <h4
-            className="font-bold text-xl cursor-pointer"
-            onClick={handleViewPostDetail}
-          >
+          <h4 className="font-bold text-xl cursor-pointer" onClick={handleViewPostDetail}>
             {post.title}
           </h4>
 
           {post.content.trim() && (
-            <ExpandableWrapper
-              maxHeight={images.length > 0 ? 100 : 500}
-              variant="text"
-            >
+            <ExpandableWrapper maxHeight={images.length > 0 ? 100 : 500} variant="text">
               <TiptapView value={post.content} />
             </ExpandableWrapper>
           )}
@@ -214,13 +200,7 @@ const PostCard = ({ post, onViewPostDetail, onDelete }: PostCardProps) => {
 
           <div className="flex items-center gap-3 mt-2">
             <Button
-              icon={
-                post.liked ? (
-                  <FaHeart className="fill-rose-600" />
-                ) : (
-                  <FaRegHeart className="fill-gray-400" />
-                )
-              }
+              icon={post.liked ? <FaHeart className="fill-rose-600" /> : <FaRegHeart className="fill-gray-400" />}
               shape="round"
               type="text"
               loading={isSubmittingLike}
